+++ app/services/contact/filter/updated_at.rb
@@ -1,0 +1,5 @@
+class Contact::Filter::UpdatedAt < Contact::Filter::Base
+  def execute_query(contacts, filters)
+    contacts.where(updated_at: filters[:updated_at])
+  end
+end